% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(n,x,y)
}
\arguments{
\item{n}{is a positive integer of bootstrap replicates.}

\item{x}{is the first variable.}

\item{y}{is the second variable. The length of y has to be the same as the length of x.}
}
\value{
Returns a list containing: correlation between the variables x and y, the bootstrap estimation of the correlation, the bootstrap estimation of the standard error, the bias, and the confidence interval.
}
\description{
Run a bootstrap for two random variables and estimate the correlation between them.
}
\examples{
bootstrap(n=1000, x=cars$speed, y=cars$dist)
}
\author{
Tobias Mortensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, Odense \cr
\email{tomor14@student.sdu.dk} \cr
}
\keyword{bootstrap}
\keyword{correlation}

