% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffons.needle.R
\name{buffons.needle}
\alias{buffons.needle}
\title{Buffons needle}
\usage{
buffons.needle(N,l,d)
}
\arguments{
\item{N}{is number of needles to be thrown. N has to be a positive integer.}

\item{l}{is the length of the needles. the length has to be positive and numeric.}

\item{d}{is the distance between the lines on the floor. The distance has to be positive and numeric.}
}
\value{
Returns the estimated value of pi by Buffons experiment.
}
\description{
Estimating pi by simulation Buffons experiment.
}
\examples{
buffons.needle(N=1000, l=1, d=1)
}
\author{
Tobias Mortensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, Odense \cr
\email{tomor14@student.sdu.dk} \cr
}
\keyword{buffon}
\keyword{needle}

