% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.goodness.R
\name{chi.goodness}
\alias{chi.goodness}
\title{Chi-squared goodness of fit}
\usage{
chi.goodness(x, p)
}
\arguments{
\item{x}{is a categorical variable. It has to be numeric}

\item{p}{is the probability of all outcomes of x. If p is not given, the outcome has uniform probability. It has to be numeric}
}
\value{
Returns the chi-squared goodness of fit value.
}
\description{
Calculate the chi-squared goodness of fit value from a random variable and a vector with the probabilities for the outcome for the variable.
}
\examples{
chi.goodness(x=c(1,3,4,1,2),p=c(.2,.3,.1,.2,.2))
}
\author{
Tobias Mortensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, Odense \cr
\email{tomor14@student.sdu.dk} \cr
}
\keyword{chi-squared}
\keyword{fit}
\keyword{goodness}
\keyword{of}

