% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.probability.R
\name{chi.probability}
\alias{chi.probability}
\title{Chi-squared probability}
\usage{
chi.probability(x, df, n)
}
\arguments{
\item{x}{is the value for which the function calculate the p-value. x has to be numeric.}

\item{df}{is the degrees of freedom by generating n number of random numbers for each variable. df has to be a positive integer.}

\item{n}{is the number af random nubers for each variable. n has to be a positive integer.}
}
\value{
Returns the p-value for the chi-squared distribution.
}
\description{
Use Monte Carlo to generate a chi-squared distribution and calculate the p-value for the distribution.
}
\examples{
chi.probability(x=15, df=19, n=10000)
}
\author{
Tobias Mortensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, Odense \cr
\email{tomor14@student.sdu.dk} \cr
}
\keyword{chi-squared}
\keyword{probability}

