% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.estimate.R
\name{density.estimate}
\alias{density.estimate}
\title{Density estimate}
\usage{
density.estimate(x, d, h, method)
}
\arguments{
\item{x}{is a numeric variable.}

\item{d}{is a point to obtain the density. d has to be a positive integer.}

\item{h}{is bandwith. h has to be a positive numeric.}

\item{method}{is either the naive estimator or the Gaussian kernel estimator. The default is the naive estimator.}
}
\value{
If d is given, it returns the density for the point d. If d is not given, it returns a list containing the density to the following points: min, 1st quartile, median, mean, 3rd qurtile and max.
}
\description{
Estimate the density of a variable, either using the naive estimator og the Gaussian kernel estimator.
}
\examples{
density.estimate(x=cars$dist, d=4, h=.8, method="naive")
}
\author{
Tobias Mortensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, Odense \cr
\email{tomor14@student.sdu.dk} \cr
}
\keyword{density}
\keyword{estimate}
\keyword{kernel}
\keyword{naive}

