% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.model.R
\name{lin.model}
\alias{lin.model}
\title{Linear regression model}
\usage{
density.plot(x, n, method, from, to)
}
\arguments{
\item{x}{is a numeric variable.}

\item{n}{is the number of points to the plot. n has to be a positive integer.}

\item{method}{is either the naive estimator or the Gaussian kernel estimator. The default is the naive estimator.}

\item{from}{is the point, you are plotting from.}

\item{to}{is the point, you are plotting to.}
}
\value{
The function gives a plot of the density, either naive or Gaussian kernel.
}
\description{
Use the density.estimate() and plot the density function.
}
\examples{
lin.model(cars$dist~cars$speed)
}
\author{
Tobias Mortensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, Odense \cr
\email{tomor14@student.sdu.dk} \cr
}
\keyword{density}
\keyword{kernel}
\keyword{naive}
\keyword{plot}

