% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.sim.R
\name{MC.sim}
\alias{MC.sim}
\title{Markov Chain simulation}
\usage{
MC.sim(p, k, n)
}
\arguments{
\item{p}{is the probability matrix for the Markov Chain. p has to be a n x n matrix.}

\item{k}{is the initial state.}

\item{n}{is number af steps to be simulated.}
}
\value{
Returns the stationary probabilities for each state.
}
\description{
Simulate the stationary discrete Markov Chain distribution.
}
\examples{
MC.sim(p=matrix(c(.6,.4,.4,.6), ncol=2), k=2, n=1000)
}
\author{
Tobias Mortensen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, Odense \cr
\email{tomor14@student.sdu.dk} \cr
}
\keyword{Chain}
\keyword{Markov}
\keyword{distribution}
\keyword{state}
\keyword{stationary}

